refactoring twitchwrapper
